requirements.txt
streamlit
streamlit_pandas_profiling
transformers
pdf2image
torch
pytesseract
tesseract
